requirements.txt
anthropic>=0.20.0
logfire>=0.29.0
pydantic>=2.5.0
requests>=2.31.0
Pillow>=10.0.0
gradio>=4.0.0
pandas>=2.0.0
pdf2image>=1.17.0
PyMuPDF>=1.23.0
prettytable>=3.9.0
seaborn>=0.12.0
matplotlib>=3.7.0
plotly>=5.17.0
python-dotenv>=1.0.0